fix: updated useBattery test and added useClipboard test
