+++ test/utils/index.ts
@@ -1,0 +1,10 @@
+import { VNode, CreateElement } from 'vue';
+
+export const mountCompositionFunc = (cb: () => any) => ({
+  setup() {
+    return cb();
+  },
+  render(h: CreateElement): VNode {
+    return h('div');
+  }
+});
